09/01/2025
spouseRoom_ now works as far as I've tested. Regular afternoon does not, function won't run for whatever reason. May be due to how I implemented it but I need to do more debugging. Also intending to patch out setUpForOutdoorPatioActivity, marriageDuties, arriveAtFarmHouse, and checkForMarriageDialogue - that should be it in the game but will plan more if needed.