09/04/2025
Found solutions to both of my problems and have written a postfix! Worked for the 5 minutes I tested it. Still need to do the Strings\\StringsFromCSFiles stuff but that's mostly busywork. Only other mandatory thing I want to sort out is to parse for dialogue formatted like {Game1.currentSeason}_{Game1.currentDay}_{spouseName} but it shouldn't really be too much of a hassle. Might try finding another way to get the spouse name for compatibility's sake but I'm satisfied with it at the moment.